backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pyyaml
openapi-spec-validator
pytest
httpx
python-dotenv
slowapi
requests
